+++ src/components/SingleValue.js
@@ -1,7 +1,6 @@
 // @flow
 import React from 'react';
 
-import { marginHorizontal } from '../mixins';
 import { Div } from '../primitives';
 import { colors, spacing } from '../theme';
 import { className } from '../utils';
@@ -11,23 +10,31 @@ type State = { isDisabled: boolean };
 type ValueProps = {
   children: string,
   data: any,
+  innerProps: any,
 };
 type Props = PropsWithStyles & ValueProps & State;
 
 export const css = ({ isDisabled }: State) => ({
-  ...marginHorizontal(spacing.baseUnit / 2),
   color: isDisabled ? colors.neutral40 : colors.text,
+  marginLeft: spacing.baseUnit / 2,
+  marginRight: spacing.baseUnit / 2,
+  maxWidth: '100%',
+  overflow: 'hidden',
   position: 'absolute',
+  textOverflow: 'ellipsis',
+  whiteSpace: 'nowrap',
 });
 
 const SingleValue = (props: Props) => {
-  const { data, getStyles, isDisabled, ...cleanProps } = props;
+  const { children, getStyles, isDisabled, innerProps } = props;
   return (
     <Div
       className={className('single-value', { isDisabled })}
       css={getStyles('singleValue', props)}
-      {...cleanProps}
-    />
+      {...innerProps}
+    >
+      {children}
+    </Div>
   );
 };