+++ src/types.js
@@ -1,11 +1,16 @@
 // @flow
-import { type ElementRef } from 'react';
+import type { Ref } from 'react';
 
 export type OptionType = {
   [string]: any,
 };
 
 export type OptionsType = Array<OptionType>;
+
+export type GroupType = {
+  options: OptionsType,
+  [string]: any,
+};
 
 export type ValueType = OptionType | OptionsType | null | void;
 
@@ -15,12 +20,25 @@ export type KeyboardEventHandler = (
   SyntheticKeyboardEvent<HTMLElement>
 ) => void;
 
-export type InnerRef = ElementRef<typeof HTMLElement>;
+export type InnerRef = Ref<*>;
 export type PropsWithInnerRef = {
-  innerRef: InnerRef,
+  innerRef: Ref<*>,
 };
+
 export type PropsWithStyles = {
   getStyles: (string, any) => {},
+};
+
+export type CommonProps = {
+  clearValue: () => void,
+  getStyles: (string, any) => {},
+  getValue: () => ValueType,
+  hasValue: boolean,
+  isMulti: boolean,
+  options: OptionsType,
+  selectOption: OptionType => void,
+  selectProps: any,
+  setValue: (ValueType, ActionTypes) => void,
 };
 
 export type ActionTypes =